I want you to scan the project to understand its goal and its development philosophy, then get to work creating the Funconium Engine and its demo.
I have placed two assets in /reources/models that I would like used in the demo when it is time launch (wont' be for ages).

For now, let's get an MVP we want to be able to create a Demo that registers controls, draws a screen, plays sound, and loads gracefully and presentably when it starts up and shuts down. Maybe give us a menu with some options and the ability to quit. Don't forget to bind ALT+F4 to quit, little details like that get lost.

Let's first make spec-first requirements for each section as we go. With hierarchical specifications, we only have to worry about one "bin" at a time and can get full context processing and output for a single directory at a time, with subdirectories getting their own turn.

However, when doing things like this, we drill-down from abstractions to concrete workings, but we need a mechanism for feedback. So whenever a directory is processesed, any problems with this part of the program OR incompatibilities with other programs (or other real-world limitations with cross-components), we need a feedback mechanism that collects those concerns. So whenever we finish a subdirectory, we will export the concerns to the parent directory to deal with. And in this way we'll keep our context inspections minimal.

So the goal right now is to flesh out the requirements for our program. We will start with C code with like structs and functions and just forget C++'s weird objects and other quirks for now. I think there are enough libraries available to accommodate us.