Ok, i ran the demo, "funconium_demo", and first of all it should get renamed to "dragon_raiders_demo".
Second, there was no window that launched with it. Let's get a window launching and showing something, some text or whatever.

Third, the command line program just printed out "Drawing the screen..." and "Playing sound..." over and over again. I don't think these should be printed out every loop so let's tight this up.

Finally, let's flesh out any part of the game engine that needs it next. We want to be able to register controls for camera updates, or do do camera shake, or otherwise animate the camera. but the cpu should only be saying, "gpu, here is the script for this" and not talk to it every other cycle or wait for some kind of feedback. Maybe service actor message passing where we can batch up a bunch of stuff in one asyncrhonous operations and have it passed to a service actor that will loop unroll and marshall data so that it can be swapped into the buffer of a dedicated hardware execution unit, of some kind (avx512 or vulkan resource, nearly anything).

But that's mostly for the future. FOr now, let's get this mvp. We probably need to create files like resources to be loaded into the engine. So we want like to be able to render an infinited plane or some kind of world. We need lighting. We need physics. We need destructible environment, voxel-style. Like when an "interactor" (say a shot from a rocket launcher or a bullet or a fire, etc, anythign capable of environment mutation) comes into contact with an area, the nearby items are checked for interactions. when a substance is destroyed it will decay through a particle system or be integrated with the environment like dust or sand.